Added --i18n-url option to build-require.js.
